workers/requirements.txt
fastapi
uvicorn
httpx
supabase
python-dotenv
pydantic
requests
moviepy
imageio[ffmpeg]
redis>=5.0
pillow
google-cloud-aiplatform
boto3
